+++ arrayInsertShift.py
@@ -31,37 +31,5 @@ def insertShiftArray(arr, val):
     return o_arr
 
 
-def test_insertShiftArray():
-    # Test case 1
-    i_arr = [1, 2, 3, 4, 5]
-    val = 6
-    expected = [1, 2, 3, 6, 4, 5]
-    o_arr = insertShiftArray(i_arr, val)
-    assert o_arr == expected, f"Expected {expected}, but got {o_arr}"
-
-    # Test case 2
-    i_arr = [1, 2, 3, 4]
-    val = 5
-    expected = [1, 2, 5, 3, 4]
-    o_arr = insertShiftArray(i_arr, val)
-    assert o_arr == expected, f"Expected {expected}, but got {o_arr}"
-
-    # Test case 3
-    i_arr = []
-    val = 1
-    expected = [1]
-    o_arr = insertShiftArray(i_arr, val)
-    assert o_arr == expected, f"Expected {expected}, but got {o_arr}"
-    
-    # Test case 4
-    i_arr = [1]
-    val = 2
-    expected = [1, 2]
-    o_arr = insertShiftArray(i_arr, val)
-    assert o_arr == expected, f"Expected {expected}, but got {o_arr}"
-
-    print("All test cases pass")
-
-test_insertShiftArray()
