+++ requirements.txt
@@ -1,0 +1,7 @@
+attrs==22.2.0
+exceptiongroup==1.1.1
+iniconfig==2.0.0
+packaging==23.0
+pluggy==1.0.0
+pytest==7.2.2
+tomli==2.0.1
